chore: Move types to composable
